add utxo input existence validation for txn

+++ apps/engine/priv/repo/migrations/20200203201059_create_utxos.exs
@@ -4,7 +4,7 @@ defmodule Engine.Repo.Migrations.CreateUtxos do
   def change do
     create table(:utxos) do
       # UTXO position information
-      add :pos, :integer
+      add :pos, :bigint
       add :blknum, :integer
       add :txindex, :integer
       add :oindex, :integer
